+++ pydbtools/__init__.py
@@ -1,2 +1,25 @@
 from .read_sql import read_sql
 from .get_athena_query_response import get_athena_query_response
+
+from .wrangler import (
+    create_temp_table,
+    read_sql_query,
+    read_sql_table,
+    create_athena_bucket,
+    describe_table,
+    get_query_columns_types,
+    get_query_execution,
+    get_work_group,
+    repair_table,
+    show_create_table,
+    start_query_execution,
+    stop_query_execution,
+    wait_query,
+    start_query_execution_and_wait,
+    create_temp_table,
+    create_temp_database,
+)
+
+import poetry_version
+
+__version__ = poetry_version.extract(source_file=__file__)