+++ pydbtools/utils.py
@@ -1,8 +1,13 @@
+from typing import Tuple
 import numpy as np
-
-from gluejobutils.s3 import s3_path_to_bucket_key, check_for_s3_file
 import os
+import re
+import sqlparse
 from s3fs import S3FileSystem
+import inspect
+import boto3
+from botocore.credentials import InstanceMetadataProvider, InstanceMetadataFetcher
+import awswrangler as wr
 
 # pydbtools will create a new a new S3 object (then delete it post read). In the first call read
 # the cache is empty but then filled. If pydbtools is called again the cache is referenced and
@@ -11,17 +16,145 @@ from s3fs import S3FileSystem
 # which S3FileSystem inherits.
 S3FileSystem.cachable = False
 
+# Get role specific path for athena output
+bucket = "mojap-athena-query-dump"
 
-def get_file(s3_path, check_exists=True):
+temp_database_name_prefix = "mojap_de_temp_"
+
+
+def get_default_args(func):
+    signature = inspect.signature(func)
+    return {
+        k: v.default
+        for k, v in signature.parameters.items()
+        if v.default is not inspect.Parameter.empty
+    }
+
+
+def check_temp_query(sql: str):
+    """
+    Checks if a query to a temporary table
+    has had __temp__ wrapped in quote marks.
+
+    Args:
+        sql (str): an SQL query
+
+    Raises:
+        ValueError
+    """
+    if re.findall(r'["|\']__temp__["|\']\.', sql.lower()):
+        raise ValueError(
+            "When querying a temporary database, __temp__ should not be wrapped in quotes"
+        )
+
+
+def clean_query(sql: str) -> str:
+    """
+    removes trailing whitespace, newlines and final
+    semicolon from sql for use with
+    sqlparse package
+
+    Args:
+        sql (str): The raw SQL query
+
+    Returns:
+        str: The cleaned SQL query
+    """
+    return " ".join(sql.splitlines()).strip().rstrip(";")
+
+
+def replace_temp_database_name_reference(sql: str, database_name: str) -> str:
+    """
+    Replaces references to the user's temp database __temp__
+    with the database_name string provided.
+
+    Args:
+        sql (str): The raw SQL query as a string
+        database_name (str): The database name to replace __temp__
+
+    Returns:
+        str: The new SQL query which is sent to Athena
+    """
+    # check query is valid and clean
+    parsed = sqlparse.parse(clean_query(sql))
+    new_query = []
+    for query in parsed:
+        check_temp_query(str(query))
+        for word in str(query).strip().split(" "):
+            if "__temp__." in word.lower():
+                word = word.lower().replace("__temp__.", f"{database_name}.")
+            new_query.append(word)
+        if ";" not in new_query[-1]:
+            last_entry = new_query[-1] + ";"
+        else:
+            last_entry = new_query[-1]
+        del new_query[-1]
+        new_query.append(last_entry)
+    return " ".join(new_query)
+
+
+def get_user_id_and_table_dir(
+    boto3_session=None, force_ec2: bool = False, region_name: str = "eu-west-1"
+) -> Tuple[str, str]:
+
+    if boto3_session is None:
+        boto3_session = get_boto_session(force_ec2=force_ec2, region_name=region_name)
+
+    sts_client = boto3_session.client("sts")
+    sts_resp = sts_client.get_caller_identity()
+    out_path = os.path.join("s3://", bucket, sts_resp["UserId"])
+    if out_path[-1] != "/":
+        out_path += "/"
+
+    return (sts_resp["UserId"], out_path)
+
+
+def get_database_name_from_userid(user_id: str) -> str:
+    unique_db_name = user_id.split(":")[-1].split("-", 1)[-1].replace("-", "_")
+    unique_db_name = temp_database_name_prefix + unique_db_name
+    return unique_db_name
+
+
+def get_boto_session(
+    force_ec2: bool = False, region_name: str = "eu-west-1",
+):
+
+    kwargs = {"region_name": region_name}
+    if force_ec2:
+        provider = InstanceMetadataProvider(
+            iam_role_fetcher=InstanceMetadataFetcher(timeout=1000, num_attempts=2)
+        )
+        creds = provider.load().get_frozen_credentials()
+        kwargs["aws_access_key_id"] = creds.access_key
+        kwargs["aws_secret_access_key"] = creds.secret_key
+        kwargs["aws_session_token"] = creds.token
+
+    return boto3.Session(**kwargs)
+
+
+def get_boto_client(
+    client_name: str,
+    boto3_session=None,
+    force_ec2: bool = False,
+    region_name: str = "eu-west-1",
+):
+
+    if boto3_session is None:
+        boto3_session = get_boto_session(force_ec2=force_ec2, region_name=region_name)
+
+    return boto3_session.client(client_name)
+
+
+def get_file(s3_path: str, check_exists: bool = True):
     """
     Returns an file using s3fs without caching objects (workaround for issue #10).
 
     s3_path: path to file in S3 e.g. s3://bucket/object/path.csv
-    check_exists: If True (default) will check for s3 file existance before returning file.
+    check_exists: If True (default) will check for s3 file existence before returning file.
     """
-    b, k = s3_path_to_bucket_key(s3_path)
+    b, k = s3_path.replace("s3://", "").split("/", 1)
     if check_exists:
-        if not check_for_s3_file(s3_path):
+        if not wr.s3.does_object_exist(s3_path):
             raise FileNotFoundError(f"File not found in S3. full path: {s3_path}")
     fs = S3FileSystem()
     f = fs.open(os.path.join(b, k), "rb")
@@ -50,7 +183,7 @@ _athena_meta_conversions = {
 
 # Two functions below stolen and altered from here:
 # https://github.com/moj-analytical-services/dataengineeringutils/blob/metadata_conformance/dataengineeringutils/pd_metadata_conformance.py
-def _pd_dtype_dict_from_metadata(athena_meta):
+def _pd_dtype_dict_from_metadata(athena_meta: list):
     """
     Convert the athena table metadata to the dtype dict that needs to be
     passed to the dtype argument of pd.read_csv. Also return list of columns that pandas needs to convert to dates/datetimes